Added dictionary PT and ES for bip-39 options
