+++ bip39WordsES.h
@@ -1,0 +1,4101 @@
+const char bipWordsES[2048][16] = {
+    "abaco",
+    "abdomen",
+    "abeja",
+    "abierto",
+    "abogado",
+    "abono",
+    "aborto",
+    "abrazo",
+    "abrir",
+    "abuelo",
+    "abuso",
+    "acabar",
+    "academia",
+    "acceso",
+    "accion",
+    "aceite",
+    "acelga",
+    "acento",
+    "aceptar",
+    "acido",
+    "aclarar",
+    "acne",
+    "acoger",
+    "acoso",
+    "activo",
+    "acto",
+    "actriz",
+    "actuar",
+    "acudir",
+    "acuerdo",
+    "acusar",
+    "adicto",
+    "admitir",
+    "adoptar",
+    "adorno",
+    "aduana",
+    "adulto",
+    "aereo",
+    "afectar",
+    "aficion",
+    "afinar",
+    "afirmar",
+    "agil",
+    "agitar",
+    "agonia",
+    "agosto",
+    "agotar",
+    "agregar",
+    "agrio",
+    "agua",
+    "agudo",
+    "aguila",
+    "aguja",
+    "ahogo",
+    "ahorro",
+    "aire",
+    "aislar",
+    "ajedrez",
+    "ajeno",
+    "ajuste",
+    "alacran",
+    "alambre",
+    "alarma",
+    "alba",
+    "album",
+    "alcalde",
+    "aldea",
+    "alegre",
+    "alejar",
+    "alerta",
+    "aleta",
+    "alfiler",
+    "alga",
+    "algodon",
+    "aliado",
+    "aliento",
+    "alivio",
+    "alma",
+    "almeja",
+    "almibar",
+    "altar",
+    "alteza",
+    "altivo",
+    "alto",
+    "altura",
+    "alumno",
+    "alzar",
+    "amable",
+    "amante",
+    "amapola",
+    "amargo",
+    "amasar",
+    "ambar",
+    "ambito",
+    "ameno",
+    "amigo",
+    "amistad",
+    "amor",
+    "amparo",
+    "amplio",
+    "ancho",
+    "anciano",
+    "ancla",
+    "andar",
+    "anden",
+    "anemia",
+    "angulo",
+    "anillo",
+    "animo",
+    "anis",
+    "anotar",
+    "antena",
+    "antiguo",
+    "antojo",
+    "anual",
+    "anular",
+    "anuncio",
+    "anadir",
+    "anejo",
+    "ano",
+    "apagar",
+    "aparato",
+    "apetito",
+    "apio",
+    "aplicar",
+    "apodo",
+    "aporte",
+    "apoyo",
+    "aprender",
+    "aprobar",
+    "apuesta",
+    "apuro",
+    "arado",
+    "arana",
+    "arar",
+    "arbitro",
+    "arbol",
+    "arbusto",
+    "archivo",
+    "arco",
+    "arder",
+    "ardilla",
+    "arduo",
+    "area",
+    "arido",
+    "aries",
+    "armonia",
+    "arnes",
+    "aroma",
+    "arpa",
+    "arpon",
+    "arreglo",
+    "arroz",
+    "arruga",
+    "arte",
+    "artista",
+    "asa",
+    "asado",
+    "asalto",
+    "ascenso",
+    "asegurar",
+    "aseo",
+    "asesor",
+    "asiento",
+    "asilo",
+    "asistir",
+    "asno",
+    "asombro",
+    "aspero",
+    "astilla",
+    "astro",
+    "astuto",
+    "asumir",
+    "asunto",
+    "atajo",
+    "ataque",
+    "atar",
+    "atento",
+    "ateo",
+    "atico",
+    "atleta",
+    "atomo",
+    "atraer",
+    "atroz",
+    "atun",
+    "audaz",
+    "audio",
+    "auge",
+    "aula",
+    "aumento",
+    "ausente",
+    "autor",
+    "aval",
+    "avance",
+    "avaro",
+    "ave",
+    "avellana",
+    "avena",
+    "avestruz",
+    "avion",
+    "aviso",
+    "ayer",
+    "ayuda",
+    "ayuno",
+    "azafran",
+    "azar",
+    "azote",
+    "azucar",
+    "azufre",
+    "azul",
+    "baba",
+    "babor",
+    "bache",
+    "bahia",
+    "baile",
+    "bajar",
+    "balanza",
+    "balcon",
+    "balde",
+    "bambu",
+    "banco",
+    "banda",
+    "bano",
+    "barba",
+    "barco",
+    "barniz",
+    "barro",
+    "bascula",
+    "baston",
+    "basura",
+    "batalla",
+    "bateria",
+    "batir",
+    "batuta",
+    "baul",
+    "bazar",
+    "bebe",
+    "bebida",
+    "bello",
+    "besar",
+    "beso",
+    "bestia",
+    "bicho",
+    "bien",
+    "bingo",
+    "blanco",
+    "bloque",
+    "blusa",
+    "boa",
+    "bobina",
+    "bobo",
+    "boca",
+    "bocina",
+    "boda",
+    "bodega",
+    "boina",
+    "bola",
+    "bolero",
+    "bolsa",
+    "bomba",
+    "bondad",
+    "bonito",
+    "bono",
+    "bonsai",
+    "borde",
+    "borrar",
+    "bosque",
+    "bote",
+    "botin",
+    "boveda",
+    "bozal",
+    "bravo",
+    "brazo",
+    "brecha",
+    "breve",
+    "brillo",
+    "brinco",
+    "brisa",
+    "broca",
+    "broma",
+    "bronce",
+    "brote",
+    "bruja",
+    "brusco",
+    "bruto",
+    "buceo",
+    "bucle",
+    "bueno",
+    "buey",
+    "bufanda",
+    "bufon",
+    "buho",
+    "buitre",
+    "bulto",
+    "burbuja",
+    "burla",
+    "burro",
+    "buscar",
+    "butaca",
+    "buzon",
+    "caballo",
+    "cabeza",
+    "cabina",
+    "cabra",
+    "cacao",
+    "cadaver",
+    "cadena",
+    "caer",
+    "cafe",
+    "caida",
+    "caiman",
+    "caja",
+    "cajon",
+    "cal",
+    "calamar",
+    "calcio",
+    "caldo",
+    "calidad",
+    "calle",
+    "calma",
+    "calor",
+    "calvo",
+    "cama",
+    "cambio",
+    "camello",
+    "camino",
+    "campo",
+    "cancer",
+    "candil",
+    "canela",
+    "canguro",
+    "canica",
+    "canto",
+    "cana",
+    "canon",
+    "caoba",
+    "caos",
+    "capaz",
+    "capitan",
+    "capote",
+    "captar",
+    "capucha",
+    "cara",
+    "carbon",
+    "carcel",
+    "careta",
+    "carga",
+    "carino",
+    "carne",
+    "carpeta",
+    "carro",
+    "carta",
+    "casa",
+    "casco",
+    "casero",
+    "caspa",
+    "castor",
+    "catorce",
+    "catre",
+    "caudal",
+    "causa",
+    "cazo",
+    "cebolla",
+    "ceder",
+    "cedro",
+    "celda",
+    "celebre",
+    "celoso",
+    "celula",
+    "cemento",
+    "ceniza",
+    "centro",
+    "cerca",
+    "cerdo",
+    "cereza",
+    "cero",
+    "cerrar",
+    "certeza",
+    "cesped",
+    "cetro",
+    "chacal",
+    "chaleco",
+    "champu",
+    "chancla",
+    "chapa",
+    "charla",
+    "chico",
+    "chiste",
+    "chivo",
+    "choque",
+    "choza",
+    "chuleta",
+    "chupar",
+    "ciclon",
+    "ciego",
+    "cielo",
+    "cien",
+    "cierto",
+    "cifra",
+    "cigarro",
+    "cima",
+    "cinco",
+    "cine",
+    "cinta",
+    "cipres",
+    "circo",
+    "ciruela",
+    "cisne",
+    "cita",
+    "ciudad",
+    "clamor",
+    "clan",
+    "claro",
+    "clase",
+    "clave",
+    "cliente",
+    "clima",
+    "clinica",
+    "cobre",
+    "coccion",
+    "cochino",
+    "cocina",
+    "coco",
+    "codigo",
+    "codo",
+    "cofre",
+    "coger",
+    "cohete",
+    "cojin",
+    "cojo",
+    "cola",
+    "colcha",
+    "colegio",
+    "colgar",
+    "colina",
+    "collar",
+    "colmo",
+    "columna",
+    "combate",
+    "comer",
+    "comida",
+    "comodo",
+    "compra",
+    "conde",
+    "conejo",
+    "conga",
+    "conocer",
+    "consejo",
+    "contar",
+    "copa",
+    "copia",
+    "corazon",
+    "corbata",
+    "corcho",
+    "cordon",
+    "corona",
+    "correr",
+    "coser",
+    "cosmos",
+    "costa",
+    "craneo",
+    "crater",
+    "crear",
+    "crecer",
+    "creido",
+    "crema",
+    "cria",
+    "crimen",
+    "cripta",
+    "crisis",
+    "cromo",
+    "cronica",
+    "croqueta",
+    "crudo",
+    "cruz",
+    "cuadro",
+    "cuarto",
+    "cuatro",
+    "cubo",
+    "cubrir",
+    "cuchara",
+    "cuello",
+    "cuento",
+    "cuerda",
+    "cuesta",
+    "cueva",
+    "cuidar",
+    "culebra",
+    "culpa",
+    "culto",
+    "cumbre",
+    "cumplir",
+    "cuna",
+    "cuneta",
+    "cuota",
+    "cupon",
+    "cupula",
+    "curar",
+    "curioso",
+    "curso",
+    "curva",
+    "cutis",
+    "dama",
+    "danza",
+    "dar",
+    "dardo",
+    "datil",
+    "deber",
+    "debil",
+    "decada",
+    "decir",
+    "dedo",
+    "defensa",
+    "definir",
+    "dejar",
+    "delfin",
+    "delgado",
+    "delito",
+    "demora",
+    "denso",
+    "dental",
+    "deporte",
+    "derecho",
+    "derrota",
+    "desayuno",
+    "deseo",
+    "desfile",
+    "desnudo",
+    "destino",
+    "desvio",
+    "detalle",
+    "detener",
+    "deuda",
+    "dia",
+    "diablo",
+    "diadema",
+    "diamante",
+    "diana",
+    "diario",
+    "dibujo",
+    "dictar",
+    "diente",
+    "dieta",
+    "diez",
+    "dificil",
+    "digno",
+    "dilema",
+    "diluir",
+    "dinero",
+    "directo",
+    "dirigir",
+    "disco",
+    "diseno",
+    "disfraz",
+    "diva",
+    "divino",
+    "doble",
+    "doce",
+    "dolor",
+    "domingo",
+    "don",
+    "donar",
+    "dorado",
+    "dormir",
+    "dorso",
+    "dos",
+    "dosis",
+    "dragon",
+    "droga",
+    "ducha",
+    "duda",
+    "duelo",
+    "dueno",
+    "dulce",
+    "duo",
+    "duque",
+    "durar",
+    "dureza",
+    "duro",
+    "ebano",
+    "ebrio",
+    "echar",
+    "eco",
+    "ecuador",
+    "edad",
+    "edicion",
+    "edificio",
+    "editor",
+    "educar",
+    "efecto",
+    "eficaz",
+    "eje",
+    "ejemplo",
+    "elefante",
+    "elegir",
+    "elemento",
+    "elevar",
+    "elipse",
+    "elite",
+    "elixir",
+    "elogio",
+    "eludir",
+    "embudo",
+    "emitir",
+    "emocion",
+    "empate",
+    "empeno",
+    "empleo",
+    "empresa",
+    "enano",
+    "encargo",
+    "enchufe",
+    "encia",
+    "enemigo",
+    "enero",
+    "enfado",
+    "enfermo",
+    "engano",
+    "enigma",
+    "enlace",
+    "enorme",
+    "enredo",
+    "ensayo",
+    "ensenar",
+    "entero",
+    "entrar",
+    "envase",
+    "envio",
+    "epoca",
+    "equipo",
+    "erizo",
+    "escala",
+    "escena",
+    "escolar",
+    "escribir",
+    "escudo",
+    "esencia",
+    "esfera",
+    "esfuerzo",
+    "espada",
+    "espejo",
+    "espia",
+    "esposa",
+    "espuma",
+    "esqui",
+    "estar",
+    "este",
+    "estilo",
+    "estufa",
+    "etapa",
+    "eterno",
+    "etica",
+    "etnia",
+    "evadir",
+    "evaluar",
+    "evento",
+    "evitar",
+    "exacto",
+    "examen",
+    "exceso",
+    "excusa",
+    "exento",
+    "exigir",
+    "exilio",
+    "existir",
+    "exito",
+    "experto",
+    "explicar",
+    "exponer",
+    "extremo",
+    "fabrica",
+    "fabula",
+    "fachada",
+    "facil",
+    "factor",
+    "faena",
+    "faja",
+    "falda",
+    "fallo",
+    "falso",
+    "faltar",
+    "fama",
+    "familia",
+    "famoso",
+    "faraon",
+    "farmacia",
+    "farol",
+    "farsa",
+    "fase",
+    "fatiga",
+    "fauna",
+    "favor",
+    "fax",
+    "febrero",
+    "fecha",
+    "feliz",
+    "feo",
+    "feria",
+    "feroz",
+    "fertil",
+    "fervor",
+    "festin",
+    "fiable",
+    "fianza",
+    "fiar",
+    "fibra",
+    "ficcion",
+    "ficha",
+    "fideo",
+    "fiebre",
+    "fiel",
+    "fiera",
+    "fiesta",
+    "figura",
+    "fijar",
+    "fijo",
+    "fila",
+    "filete",
+    "filial",
+    "filtro",
+    "fin",
+    "finca",
+    "fingir",
+    "finito",
+    "firma",
+    "flaco",
+    "flauta",
+    "flecha",
+    "flor",
+    "flota",
+    "fluir",
+    "flujo",
+    "fluor",
+    "fobia",
+    "foca",
+    "fogata",
+    "fogon",
+    "folio",
+    "folleto",
+    "fondo",
+    "forma",
+    "forro",
+    "fortuna",
+    "forzar",
+    "fosa",
+    "foto",
+    "fracaso",
+    "fragil",
+    "franja",
+    "frase",
+    "fraude",
+    "freir",
+    "freno",
+    "fresa",
+    "frio",
+    "frito",
+    "fruta",
+    "fuego",
+    "fuente",
+    "fuerza",
+    "fuga",
+    "fumar",
+    "funcion",
+    "funda",
+    "furgon",
+    "furia",
+    "fusil",
+    "futbol",
+    "futuro",
+    "gacela",
+    "gafas",
+    "gaita",
+    "gajo",
+    "gala",
+    "galeria",
+    "gallo",
+    "gamba",
+    "ganar",
+    "gancho",
+    "ganga",
+    "ganso",
+    "garaje",
+    "garza",
+    "gasolina",
+    "gastar",
+    "gato",
+    "gavilan",
+    "gemelo",
+    "gemir",
+    "gen",
+    "genero",
+    "genio",
+    "gente",
+    "geranio",
+    "gerente",
+    "germen",
+    "gesto",
+    "gigante",
+    "gimnasio",
+    "girar",
+    "giro",
+    "glaciar",
+    "globo",
+    "gloria",
+    "gol",
+    "golfo",
+    "goloso",
+    "golpe",
+    "goma",
+    "gordo",
+    "gorila",
+    "gorra",
+    "gota",
+    "goteo",
+    "gozar",
+    "grada",
+    "grafico",
+    "grano",
+    "grasa",
+    "gratis",
+    "grave",
+    "grieta",
+    "grillo",
+    "gripe",
+    "gris",
+    "grito",
+    "grosor",
+    "grua",
+    "grueso",
+    "grumo",
+    "grupo",
+    "guante",
+    "guapo",
+    "guardia",
+    "guerra",
+    "guia",
+    "guino",
+    "guion",
+    "guiso",
+    "guitarra",
+    "gusano",
+    "gustar",
+    "haber",
+    "habil",
+    "hablar",
+    "hacer",
+    "hacha",
+    "hada",
+    "hallar",
+    "hamaca",
+    "harina",
+    "haz",
+    "hazana",
+    "hebilla",
+    "hebra",
+    "hecho",
+    "helado",
+    "helio",
+    "hembra",
+    "herir",
+    "hermano",
+    "heroe",
+    "hervir",
+    "hielo",
+    "hierro",
+    "higado",
+    "higiene",
+    "hijo",
+    "himno",
+    "historia",
+    "hocico",
+    "hogar",
+    "hoguera",
+    "hoja",
+    "hombre",
+    "hongo",
+    "honor",
+    "honra",
+    "hora",
+    "hormiga",
+    "horno",
+    "hostil",
+    "hoyo",
+    "hueco",
+    "huelga",
+    "huerta",
+    "hueso",
+    "huevo",
+    "huida",
+    "huir",
+    "humano",
+    "humedo",
+    "humilde",
+    "humo",
+    "hundir",
+    "huracan",
+    "hurto",
+    "icono",
+    "ideal",
+    "idioma",
+    "idolo",
+    "iglesia",
+    "iglu",
+    "igual",
+    "ilegal",
+    "ilusion",
+    "imagen",
+    "iman",
+    "imitar",
+    "impar",
+    "imperio",
+    "imponer",
+    "impulso",
+    "incapaz",
+    "indice",
+    "inerte",
+    "infiel",
+    "informe",
+    "ingenio",
+    "inicio",
+    "inmenso",
+    "inmune",
+    "innato",
+    "insecto",
+    "instante",
+    "interes",
+    "intimo",
+    "intuir",
+    "inutil",
+    "invierno",
+    "ira",
+    "iris",
+    "ironia",
+    "isla",
+    "islote",
+    "jabali",
+    "jabon",
+    "jamon",
+    "jarabe",
+    "jardin",
+    "jarra",
+    "jaula",
+    "jazmin",
+    "jefe",
+    "jeringa",
+    "jinete",
+    "jornada",
+    "joroba",
+    "joven",
+    "joya",
+    "juerga",
+    "jueves",
+    "juez",
+    "jugador",
+    "jugo",
+    "juguete",
+    "juicio",
+    "junco",
+    "jungla",
+    "junio",
+    "juntar",
+    "jupiter",
+    "jurar",
+    "justo",
+    "juvenil",
+    "juzgar",
+    "kilo",
+    "koala",
+    "labio",
+    "lacio",
+    "lacra",
+    "lado",
+    "ladron",
+    "lagarto",
+    "lagrima",
+    "laguna",
+    "laico",
+    "lamer",
+    "lamina",
+    "lampara",
+    "lana",
+    "lancha",
+    "langosta",
+    "lanza",
+    "lapiz",
+    "largo",
+    "larva",
+    "lastima",
+    "lata",
+    "latex",
+    "latir",
+    "laurel",
+    "lavar",
+    "lazo",
+    "leal",
+    "leccion",
+    "leche",
+    "lector",
+    "leer",
+    "legion",
+    "legumbre",
+    "lejano",
+    "lengua",
+    "lento",
+    "lena",
+    "leon",
+    "leopardo",
+    "lesion",
+    "letal",
+    "letra",
+    "leve",
+    "leyenda",
+    "libertad",
+    "libro",
+    "licor",
+    "lider",
+    "lidiar",
+    "lienzo",
+    "liga",
+    "ligero",
+    "lima",
+    "limite",
+    "limon",
+    "limpio",
+    "lince",
+    "lindo",
+    "linea",
+    "lingote",
+    "lino",
+    "linterna",
+    "liquido",
+    "liso",
+    "lista",
+    "litera",
+    "litio",
+    "litro",
+    "llaga",
+    "llama",
+    "llanto",
+    "llave",
+    "llegar",
+    "llenar",
+    "llevar",
+    "llorar",
+    "llover",
+    "lluvia",
+    "lobo",
+    "locion",
+    "loco",
+    "locura",
+    "logica",
+    "logro",
+    "lombriz",
+    "lomo",
+    "lonja",
+    "lote",
+    "lucha",
+    "lucir",
+    "lugar",
+    "lujo",
+    "luna",
+    "lunes",
+    "lupa",
+    "lustro",
+    "luto",
+    "luz",
+    "maceta",
+    "macho",
+    "madera",
+    "madre",
+    "maduro",
+    "maestro",
+    "mafia",
+    "magia",
+    "mago",
+    "maiz",
+    "maldad",
+    "maleta",
+    "malla",
+    "malo",
+    "mama",
+    "mambo",
+    "mamut",
+    "manco",
+    "mando",
+    "manejar",
+    "manga",
+    "maniqui",
+    "manjar",
+    "mano",
+    "manso",
+    "manta",
+    "manana",
+    "mapa",
+    "maquina",
+    "mar",
+    "marco",
+    "marea",
+    "marfil",
+    "margen",
+    "marido",
+    "marmol",
+    "marron",
+    "martes",
+    "marzo",
+    "masa",
+    "mascara",
+    "masivo",
+    "matar",
+    "materia",
+    "matiz",
+    "matriz",
+    "maximo",
+    "mayor",
+    "mazorca",
+    "mecha",
+    "medalla",
+    "medio",
+    "medula",
+    "mejilla",
+    "mejor",
+    "melena",
+    "melon",
+    "memoria",
+    "menor",
+    "mensaje",
+    "mente",
+    "menu",
+    "mercado",
+    "merengue",
+    "merito",
+    "mes",
+    "meson",
+    "meta",
+    "meter",
+    "metodo",
+    "metro",
+    "mezcla",
+    "miedo",
+    "miel",
+    "miembro",
+    "miga",
+    "mil",
+    "milagro",
+    "militar",
+    "millon",
+    "mimo",
+    "mina",
+    "minero",
+    "minimo",
+    "minuto",
+    "miope",
+    "mirar",
+    "misa",
+    "miseria",
+    "misil",
+    "mismo",
+    "mitad",
+    "mito",
+    "mochila",
+    "mocion",
+    "moda",
+    "modelo",
+    "moho",
+    "mojar",
+    "molde",
+    "moler",
+    "molino",
+    "momento",
+    "momia",
+    "monarca",
+    "moneda",
+    "monja",
+    "monto",
+    "mono",
+    "morada",
+    "morder",
+    "moreno",
+    "morir",
+    "morro",
+    "morsa",
+    "mortal",
+    "mosca",
+    "mostrar",
+    "motivo",
+    "mover",
+    "movil",
+    "mozo",
+    "mucho",
+    "mudar",
+    "mueble",
+    "muela",
+    "muerte",
+    "muestra",
+    "mugre",
+    "mujer",
+    "mula",
+    "muleta",
+    "multa",
+    "mundo",
+    "muneca",
+    "mural",
+    "muro",
+    "musculo",
+    "museo",
+    "musgo",
+    "musica",
+    "muslo",
+    "nacar",
+    "nacion",
+    "nadar",
+    "naipe",
+    "naranja",
+    "nariz",
+    "narrar",
+    "nasal",
+    "natal",
+    "nativo",
+    "natural",
+    "nausea",
+    "naval",
+    "nave",
+    "navidad",
+    "necio",
+    "nectar",
+    "negar",
+    "negocio",
+    "negro",
+    "neon",
+    "nervio",
+    "neto",
+    "neutro",
+    "nevar",
+    "nevera",
+    "nicho",
+    "nido",
+    "niebla",
+    "nieto",
+    "ninez",
+    "nino",
+    "nitido",
+    "nivel",
+    "nobleza",
+    "noche",
+    "nomina",
+    "noria",
+    "norma",
+    "norte",
+    "nota",
+    "noticia",
+    "novato",
+    "novela",
+    "novio",
+    "nube",
+    "nuca",
+    "nucleo",
+    "nudillo",
+    "nudo",
+    "nuera",
+    "nueve",
+    "nuez",
+    "nulo",
+    "numero",
+    "nutria",
+    "oasis",
+    "obeso",
+    "obispo",
+    "objeto",
+    "obra",
+    "obrero",
+    "observar",
+    "obtener",
+    "obvio",
+    "oca",
+    "ocaso",
+    "oceano",
+    "ochenta",
+    "ocho",
+    "ocio",
+    "ocre",
+    "octavo",
+    "octubre",
+    "oculto",
+    "ocupar",
+    "ocurrir",
+    "odiar",
+    "odio",
+    "odisea",
+    "oeste",
+    "ofensa",
+    "oferta",
+    "oficio",
+    "ofrecer",
+    "ogro",
+    "oido",
+    "oir",
+    "ojo",
+    "ola",
+    "oleada",
+    "olfato",
+    "olivo",
+    "olla",
+    "olmo",
+    "olor",
+    "olvido",
+    "ombligo",
+    "onda",
+    "onza",
+    "opaco",
+    "opcion",
+    "opera",
+    "opinar",
+    "oponer",
+    "optar",
+    "optica",
+    "opuesto",
+    "oracion",
+    "orador",
+    "oral",
+    "orbita",
+    "orca",
+    "orden",
+    "oreja",
+    "organo",
+    "orgia",
+    "orgullo",
+    "oriente",
+    "origen",
+    "orilla",
+    "oro",
+    "orquesta",
+    "oruga",
+    "osadia",
+    "oscuro",
+    "osezno",
+    "oso",
+    "ostra",
+    "otono",
+    "otro",
+    "oveja",
+    "ovulo",
+    "oxido",
+    "oxigeno",
+    "oyente",
+    "ozono",
+    "pacto",
+    "padre",
+    "paella",
+    "pagina",
+    "pago",
+    "pais",
+    "pajaro",
+    "palabra",
+    "palco",
+    "paleta",
+    "palido",
+    "palma",
+    "paloma",
+    "palpar",
+    "pan",
+    "panal",
+    "panico",
+    "pantera",
+    "panuelo",
+    "papa",
+    "papel",
+    "papilla",
+    "paquete",
+    "parar",
+    "parcela",
+    "pared",
+    "parir",
+    "paro",
+    "parpado",
+    "parque",
+    "parrafo",
+    "parte",
+    "pasar",
+    "paseo",
+    "pasion",
+    "paso",
+    "pasta",
+    "pata",
+    "patio",
+    "patria",
+    "pausa",
+    "pauta",
+    "pavo",
+    "payaso",
+    "peaton",
+    "pecado",
+    "pecera",
+    "pecho",
+    "pedal",
+    "pedir",
+    "pegar",
+    "peine",
+    "pelar",
+    "peldano",
+    "pelea",
+    "peligro",
+    "pellejo",
+    "pelo",
+    "peluca",
+    "pena",
+    "pensar",
+    "penon",
+    "peon",
+    "peor",
+    "pepino",
+    "pequeno",
+    "pera",
+    "percha",
+    "perder",
+    "pereza",
+    "perfil",
+    "perico",
+    "perla",
+    "permiso",
+    "perro",
+    "persona",
+    "pesa",
+    "pesca",
+    "pesimo",
+    "pestana",
+    "petalo",
+    "petroleo",
+    "pez",
+    "pezuna",
+    "picar",
+    "pichon",
+    "pie",
+    "piedra",
+    "pierna",
+    "pieza",
+    "pijama",
+    "pilar",
+    "piloto",
+    "pimienta",
+    "pino",
+    "pintor",
+    "pinza",
+    "pina",
+    "piojo",
+    "pipa",
+    "pirata",
+    "pisar",
+    "piscina",
+    "piso",
+    "pista",
+    "piton",
+    "pizca",
+    "placa",
+    "plan",
+    "plata",
+    "playa",
+    "plaza",
+    "pleito",
+    "pleno",
+    "plomo",
+    "pluma",
+    "plural",
+    "pobre",
+    "poco",
+    "poder",
+    "podio",
+    "poema",
+    "poesia",
+    "poeta",
+    "polen",
+    "policia",
+    "pollo",
+    "polvo",
+    "pomada",
+    "pomelo",
+    "pomo",
+    "pompa",
+    "poner",
+    "porcion",
+    "portal",
+    "posada",
+    "poseer",
+    "posible",
+    "poste",
+    "potencia",
+    "potro",
+    "pozo",
+    "prado",
+    "precoz",
+    "pregunta",
+    "premio",
+    "prensa",
+    "preso",
+    "previo",
+    "primo",
+    "principe",
+    "prision",
+    "privar",
+    "proa",
+    "probar",
+    "proceso",
+    "producto",
+    "proeza",
+    "profesor",
+    "programa",
+    "prole",
+    "promesa",
+    "pronto",
+    "propio",
+    "proximo",
+    "prueba",
+    "publico",
+    "puchero",
+    "pudor",
+    "pueblo",
+    "puerta",
+    "puesto",
+    "pulga",
+    "pulir",
+    "pulmon",
+    "pulpo",
+    "pulso",
+    "puma",
+    "punto",
+    "punal",
+    "puno",
+    "pupa",
+    "pupila",
+    "pure",
+    "quedar",
+    "queja",
+    "quemar",
+    "querer",
+    "queso",
+    "quieto",
+    "quimica",
+    "quince",
+    "quitar",
+    "rabano",
+    "rabia",
+    "rabo",
+    "racion",
+    "radical",
+    "raiz",
+    "rama",
+    "rampa",
+    "rancho",
+    "rango",
+    "rapaz",
+    "rapido",
+    "rapto",
+    "rasgo",
+    "raspa",
+    "rato",
+    "rayo",
+    "raza",
+    "razon",
+    "reaccion",
+    "realidad",
+    "rebano",
+    "rebote",
+    "recaer",
+    "receta",
+    "rechazo",
+    "recoger",
+    "recreo",
+    "recto",
+    "recurso",
+    "red",
+    "redondo",
+    "reducir",
+    "reflejo",
+    "reforma",
+    "refran",
+    "refugio",
+    "regalo",
+    "regir",
+    "regla",
+    "regreso",
+    "rehen",
+    "reino",
+    "reir",
+    "reja",
+    "relato",
+    "relevo",
+    "relieve",
+    "relleno",
+    "reloj",
+    "remar",
+    "remedio",
+    "remo",
+    "rencor",
+    "rendir",
+    "renta",
+    "reparto",
+    "repetir",
+    "reposo",
+    "reptil",
+    "res",
+    "rescate",
+    "resina",
+    "respeto",
+    "resto",
+    "resumen",
+    "retiro",
+    "retorno",
+    "retrato",
+    "reunir",
+    "reves",
+    "revista",
+    "rey",
+    "rezar",
+    "rico",
+    "riego",
+    "rienda",
+    "riesgo",
+    "rifa",
+    "rigido",
+    "rigor",
+    "rincon",
+    "rinon",
+    "rio",
+    "riqueza",
+    "risa",
+    "ritmo",
+    "rito",
+    "rizo",
+    "roble",
+    "roce",
+    "rociar",
+    "rodar",
+    "rodeo",
+    "rodilla",
+    "roer",
+    "rojizo",
+    "rojo",
+    "romero",
+    "romper",
+    "ron",
+    "ronco",
+    "ronda",
+    "ropa",
+    "ropero",
+    "rosa",
+    "rosca",
+    "rostro",
+    "rotar",
+    "rubi",
+    "rubor",
+    "rudo",
+    "rueda",
+    "rugir",
+    "ruido",
+    "ruina",
+    "ruleta",
+    "rulo",
+    "rumbo",
+    "rumor",
+    "ruptura",
+    "ruta",
+    "rutina",
+    "sabado",
+    "saber",
+    "sabio",
+    "sable",
+    "sacar",
+    "sagaz",
+    "sagrado",
+    "sala",
+    "saldo",
+    "salero",
+    "salir",
+    "salmon",
+    "salon",
+    "salsa",
+    "salto",
+    "salud",
+    "salvar",
+    "samba",
+    "sancion",
+    "sandia",
+    "sanear",
+    "sangre",
+    "sanidad",
+    "sano",
+    "santo",
+    "sapo",
+    "saque",
+    "sardina",
+    "sarten",
+    "sastre",
+    "satan",
+    "sauna",
+    "saxofon",
+    "seccion",
+    "seco",
+    "secreto",
+    "secta",
+    "sed",
+    "seguir",
+    "seis",
+    "sello",
+    "selva",
+    "semana",
+    "semilla",
+    "senda",
+    "sensor",
+    "senal",
+    "senor",
+    "separar",
+    "sepia",
+    "sequia",
+    "ser",
+    "serie",
+    "sermon",
+    "servir",
+    "sesenta",
+    "sesion",
+    "seta",
+    "setenta",
+    "severo",
+    "sexo",
+    "sexto",
+    "sidra",
+    "siesta",
+    "siete",
+    "siglo",
+    "signo",
+    "silaba",
+    "silbar",
+    "silencio",
+    "silla",
+    "simbolo",
+    "simio",
+    "sirena",
+    "sistema",
+    "sitio",
+    "situar",
+    "sobre",
+    "socio",
+    "sodio",
+    "sol",
+    "solapa",
+    "soldado",
+    "soledad",
+    "solido",
+    "soltar",
+    "solucion",
+    "sombra",
+    "sondeo",
+    "sonido",
+    "sonoro",
+    "sonrisa",
+    "sopa",
+    "soplar",
+    "soporte",
+    "sordo",
+    "sorpresa",
+    "sorteo",
+    "sosten",
+    "sotano",
+    "suave",
+    "subir",
+    "suceso",
+    "sudor",
+    "suegra",
+    "suelo",
+    "sueno",
+    "suerte",
+    "sufrir",
+    "sujeto",
+    "sultan",
+    "sumar",
+    "superar",
+    "suplir",
+    "suponer",
+    "supremo",
+    "sur",
+    "surco",
+    "sureno",
+    "surgir",
+    "susto",
+    "sutil",
+    "tabaco",
+    "tabique",
+    "tabla",
+    "tabu",
+    "taco",
+    "tacto",
+    "tajo",
+    "talar",
+    "talco",
+    "talento",
+    "talla",
+    "talon",
+    "tamano",
+    "tambor",
+    "tango",
+    "tanque",
+    "tapa",
+    "tapete",
+    "tapia",
+    "tapon",
+    "taquilla",
+    "tarde",
+    "tarea",
+    "tarifa",
+    "tarjeta",
+    "tarot",
+    "tarro",
+    "tarta",
+    "tatuaje",
+    "tauro",
+    "taza",
+    "tazon",
+    "teatro",
+    "techo",
+    "tecla",
+    "tecnica",
+    "tejado",
+    "tejer",
+    "tejido",
+    "tela",
+    "telefono",
+    "tema",
+    "temor",
+    "templo",
+    "tenaz",
+    "tender",
+    "tener",
+    "tenis",
+    "tenso",
+    "teoria",
+    "terapia",
+    "terco",
+    "termino",
+    "ternura",
+    "terror",
+    "tesis",
+    "tesoro",
+    "testigo",
+    "tetera",
+    "texto",
+    "tez",
+    "tibio",
+    "tiburon",
+    "tiempo",
+    "tienda",
+    "tierra",
+    "tieso",
+    "tigre",
+    "tijera",
+    "tilde",
+    "timbre",
+    "timido",
+    "timo",
+    "tinta",
+    "tio",
+    "tipico",
+    "tipo",
+    "tira",
+    "tiron",
+    "titan",
+    "titere",
+    "titulo",
+    "tiza",
+    "toalla",
+    "tobillo",
+    "tocar",
+    "tocino",
+    "todo",
+    "toga",
+    "toldo",
+    "tomar",
+    "tono",
+    "tonto",
+    "topar",
+    "tope",
+    "toque",
+    "torax",
+    "torero",
+    "tormenta",
+    "torneo",
+    "toro",
+    "torpedo",
+    "torre",
+    "torso",
+    "tortuga",
+    "tos",
+    "tosco",
+    "toser",
+    "toxico",
+    "trabajo",
+    "tractor",
+    "traer",
+    "trafico",
+    "trago",
+    "traje",
+    "tramo",
+    "trance",
+    "trato",
+    "trauma",
+    "trazar",
+    "trebol",
+    "tregua",
+    "treinta",
+    "tren",
+    "trepar",
+    "tres",
+    "tribu",
+    "trigo",
+    "tripa",
+    "triste",
+    "triunfo",
+    "trofeo",
+    "trompa",
+    "tronco",
+    "tropa",
+    "trote",
+    "trozo",
+    "truco",
+    "trueno",
+    "trufa",
+    "tuberia",
+    "tubo",
+    "tuerto",
+    "tumba",
+    "tumor",
+    "tunel",
+    "tunica",
+    "turbina",
+    "turismo",
+    "turno",
+    "tutor",
+    "ubicar",
+    "ulcera",
+    "umbral",
+    "unidad",
+    "unir",
+    "universo",
+    "uno",
+    "untar",
+    "una",
+    "urbano",
+    "urbe",
+    "urgente",
+    "urna",
+    "usar",
+    "usuario",
+    "util",
+    "utopia",
+    "uva",
+    "vaca",
+    "vacio",
+    "vacuna",
+    "vagar",
+    "vago",
+    "vaina",
+    "vajilla",
+    "vale",
+    "valido",
+    "valle",
+    "valor",
+    "valvula",
+    "vampiro",
+    "vara",
+    "variar",
+    "varon",
+    "vaso",
+    "vecino",
+    "vector",
+    "vehiculo",
+    "veinte",
+    "vejez",
+    "vela",
+    "velero",
+    "veloz",
+    "vena",
+    "vencer",
+    "venda",
+    "veneno",
+    "vengar",
+    "venir",
+    "venta",
+    "venus",
+    "ver",
+    "verano",
+    "verbo",
+    "verde",
+    "vereda",
+    "verja",
+    "verso",
+    "verter",
+    "via",
+    "viaje",
+    "vibrar",
+    "vicio",
+    "victima",
+    "vida",
+    "video",
+    "vidrio",
+    "viejo",
+    "viernes",
+    "vigor",
+    "vil",
+    "villa",
+    "vinagre",
+    "vino",
+    "vinedo",
+    "violin",
+    "viral",
+    "virgo",
+    "virtud",
+    "visor",
+    "vispera",
+    "vista",
+    "vitamina",
+    "viudo",
+    "vivaz",
+    "vivero",
+    "vivir",
+    "vivo",
+    "volcan",
+    "volumen",
+    "volver",
+    "voraz",
+    "votar",
+    "voto",
+    "voz",
+    "vuelo",
+    "vulgar",
+    "yacer",
+    "yate",
+    "yegua",
+    "yema",
+    "yerno",
+    "yeso",
+    "yodo",
+    "yoga",
+    "yogur",
+    "zafiro",
+    "zanja",
+    "zapato",
+    "zarza",
+    "zona",
+    "zorro",
+    "zumo",
+    "zurdo"
+};
+
+const char bipOffsetES[2048] = {
+    0,
+    -2,
+    2,
+    -2,
+    0,
+    2,
+    -1,
+    0,
+    1,
+    -1,
+    1,
+    -1,
+    -2,
+    2,
+    0,
+    0,
+    0,
+    0,
+    -1,
+    2,
+    -2,
+    3,
+    -2,
+    1,
+    -1,
+    2,
+    -2,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    -1,
+    0,
+    1,
+    0,
+    0,
+    1,
+    -2,
+    0,
+    1,
+    -1,
+    3,
+    -2,
+    0,
+    0,
+    0,
+    -1,
+    2,
+    1,
+    -1,
+    -1,
+    1,
+    0,
+    -1,
+    2,
+    -2,
+    -1,
+    2,
+    -1,
+    -1,
+    0,
+    1,
+    2,
+    -1,
+    -2,
+    2,
+    -1,
+    0,
+    0,
+    1,
+    -2,
+    3,
+    -3,
+    1,
+    -1,
+    1,
+    2,
+    -2,
+    -1,
+    2,
+    -1,
+    0,
+    2,
+    -2,
+    0,
+    1,
+    -1,
+    0,
+    -1,
+    1,
+    0,
+    1,
+    -1,
+    1,
+    0,
+    -2,
+    3,
+    -2,
+    0,
+    1,
+    -2,
+    2,
+    0,
+    0,
+    -1,
+    0,
+    0,
+    1,
+    1,
+    -2,
+    0,
+    -1,
+    1,
+    1,
+    -1,
+    -1,
+    1,
+    1,
+    2,
+    -3,
+    -1,
+    0,
+    3,
+    -3,
+    2,
+    -1,
+    1,
+    -3,
+    1,
+    0,
+    2,
+    0,
+    0,
+    1,
+    -3,
+    2,
+    -2,
+    0,
+    3,
+    -1,
+    -2,
+    2,
+    1,
+    -1,
+    0,
+    -2,
+    2,
+    0,
+    1,
+    -1,
+    -2,
+    2,
+    -1,
+    2,
+    -3,
+    4,
+    -2,
+    -1,
+    -1,
+    -1,
+    4,
+    -2,
+    -1,
+    2,
+    -2,
+    3,
+    -3,
+    1,
+    -1,
+    2,
+    -1,
+    0,
+    0,
+    1,
+    -1,
+    2,
+    -2,
+    2,
+    -1,
+    -1,
+    1,
+    -1,
+    1,
+    1,
+    -1,
+    0,
+    1,
+    0,
+    -3,
+    0,
+    2,
+    1,
+    -2,
+    1,
+    2,
+    -5,
+    3,
+    -3,
+    3,
+    0,
+    1,
+    -1,
+    0,
+    -2,
+    3,
+    -1,
+    -1,
+    0,
+    2,
+    0,
+    -1,
+    0,
+    0,
+    0,
+    0,
+    -2,
+    1,
+    1,
+    0,
+    0,
+    0,
+    1,
+    -1,
+    0,
+    -1,
+    1,
+    -2,
+    1,
+    0,
+    -1,
+    0,
+    2,
+    -1,
+    2,
+    -1,
+    1,
+    -2,
+    1,
+    0,
+    1,
+    -2,
+    1,
+    1,
+    -1,
+    -1,
+    0,
+    1,
+    2,
+    -3,
+    2,
+    0,
+    -2,
+    2,
+    -2,
+    1,
+    1,
+    -2,
+    1,
+    0,
+    -1,
+    0,
+    2,
+    -2,
+    1,
+    -1,
+    0,
+    2,
+    -1,
+    -1,
+    1,
+    0,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    1,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    0,
+    0,
+    1,
+    -3,
+    2,
+    1,
+    -2,
+    1,
+    -2,
+    2,
+    0,
+    -1,
+    0,
+    1,
+    -2,
+    1,
+    0,
+    1,
+    0,
+    -2,
+    1,
+    2,
+    0,
+    -1,
+    -1,
+    2,
+    -1,
+    2,
+    -4,
+    1,
+    1,
+    -2,
+    2,
+    0,
+    0,
+    0,
+    1,
+    -2,
+    -1,
+    1,
+    1,
+    -1,
+    0,
+    0,
+    -1,
+    1,
+    1,
+    1,
+    -1,
+    0,
+    1,
+    -1,
+    -2,
+    1,
+    0,
+    -1,
+    3,
+    -2,
+    0,
+    0,
+    1,
+    -1,
+    1,
+    -2,
+    2,
+    0,
+    1,
+    -1,
+    -1,
+    1,
+    -1,
+    -1,
+    2,
+    -1,
+    1,
+    1,
+    -3,
+    2,
+    0,
+    0,
+    -2,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    1,
+    0,
+    -1,
+    2,
+    -2,
+    -1,
+    1,
+    1,
+    -1,
+    -1,
+    1,
+    -1,
+    2,
+    -1,
+    1,
+    -1,
+    1,
+    -1,
+    1,
+    -2,
+    1,
+    0,
+    1,
+    0,
+    1,
+    -2,
+    1,
+    -2,
+    3,
+    -1,
+    1,
+    -1,
+    -1,
+    1,
+    -2,
+    2,
+    1,
+    -2,
+    0,
+    2,
+    -1,
+    0,
+    0,
+    -2,
+    2,
+    -2,
+    2,
+    -2,
+    0,
+    1,
+    2,
+    -2,
+    2,
+    -1,
+    0,
+    -1,
+    1,
+    1,
+    0,
+    -2,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    -2,
+    0,
+    2,
+    -1,
+    0,
+    0,
+    1,
+    -1,
+    1,
+    -2,
+    0,
+    1,
+    2,
+    -1,
+    -2,
+    0,
+    1,
+    0,
+    0,
+    0,
+    1,
+    -1,
+    1,
+    -1,
+    0,
+    1,
+    -1,
+    0,
+    1,
+    1,
+    -2,
+    0,
+    0,
+    1,
+    -2,
+    -1,
+    3,
+    1,
+    -2,
+    0,
+    0,
+    2,
+    -2,
+    -1,
+    1,
+    0,
+    0,
+    0,
+    1,
+    -1,
+    -1,
+    2,
+    0,
+    -1,
+    -1,
+    3,
+    -2,
+    1,
+    0,
+    -1,
+    1,
+    -2,
+    2,
+    0,
+    0,
+    1,
+    -1,
+    2,
+    -2,
+    0,
+    0,
+    0,
+    -1,
+    1,
+    1,
+    -3,
+    0,
+    2,
+    -1,
+    -1,
+    1,
+    0,
+    1,
+    -1,
+    -1,
+    0,
+    0,
+    -1,
+    3,
+    -2,
+    0,
+    0,
+    1,
+    -1,
+    0,
+    2,
+    2,
+    -3,
+    -1,
+    -1,
+    3,
+    -1,
+    0,
+    0,
+    0,
+    1,
+    1,
+    -3,
+    2,
+    -1,
+    0,
+    0,
+    -1,
+    0,
+    2,
+    -1,
+    -1,
+    3,
+    -2,
+    1,
+    1,
+    -1,
+    -2,
+    4,
+    -2,
+    -1,
+    0,
+    1,
+    2,
+    -2,
+    -1,
+    1,
+    0,
+    1,
+    -1,
+    0,
+    0,
+    2,
+    -2,
+    0,
+    -1,
+    2,
+    -1,
+    0,
+    0,
+    2,
+    -4,
+    3,
+    -3,
+    -1,
+    2,
+    0,
+    0,
+    0,
+    3,
+    -4,
+    -1,
+    2,
+    -2,
+    2,
+    0,
+    1,
+    -1,
+    0,
+    0,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    0,
+    -1,
+    2,
+    -2,
+    0,
+    2,
+    -2,
+    2,
+    -1,
+    -1,
+    1,
+    0,
+    0,
+    0,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    -1,
+    -1,
+    2,
+    -1,
+    1,
+    -2,
+    2,
+    0,
+    1,
+    -1,
+    0,
+    1,
+    0,
+    1,
+    -2,
+    0,
+    1,
+    -1,
+    1,
+    0,
+    -1,
+    -1,
+    1,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    -1,
+    2,
+    -2,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    -1,
+    2,
+    -1,
+    1,
+    1,
+    -1,
+    0,
+    0,
+    -1,
+    2,
+    -3,
+    1,
+    0,
+    -2,
+    3,
+    0,
+    1,
+    -2,
+    1,
+    0,
+    2,
+    -4,
+    2,
+    0,
+    2,
+    -2,
+    0,
+    -1,
+    0,
+    0,
+    0,
+    0,
+    2,
+    -1,
+    -2,
+    2,
+    0,
+    -1,
+    2,
+    -1,
+    -1,
+    0,
+    1,
+    1,
+    0,
+    -2,
+    0,
+    0,
+    3,
+    -2,
+    -1,
+    0,
+    1,
+    0,
+    -1,
+    0,
+    2,
+    -1,
+    0,
+    0,
+    0,
+    0,
+    1,
+    -2,
+    1,
+    0,
+    -2,
+    2,
+    0,
+    0,
+    -2,
+    1,
+    2,
+    0,
+    -3,
+    1,
+    0,
+    1,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    -1,
+    0,
+    0,
+    -1,
+    0,
+    2,
+    -1,
+    -2,
+    2,
+    -1,
+    1,
+    0,
+    -1,
+    0,
+    0,
+    1,
+    0,
+    1,
+    0,
+    -3,
+    2,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    -1,
+    1,
+    -3,
+    2,
+    2,
+    -3,
+    1,
+    1,
+    2,
+    -3,
+    1,
+    0,
+    -2,
+    0,
+    1,
+    1,
+    -2,
+    -1,
+    3,
+    1,
+    -3,
+    2,
+    -1,
+    3,
+    -2,
+    -1,
+    1,
+    1,
+    -1,
+    -1,
+    1,
+    1,
+    -1,
+    0,
+    0,
+    -2,
+    2,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    1,
+    1,
+    -1,
+    -1,
+    2,
+    -2,
+    1,
+    0,
+    -1,
+    1,
+    -2,
+    1,
+    2,
+    -1,
+    0,
+    0,
+    -3,
+    2,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    1,
+    -2,
+    0,
+    0,
+    3,
+    -3,
+    -1,
+    2,
+    0,
+    -1,
+    1,
+    -1,
+    1,
+    -2,
+    2,
+    -1,
+    1,
+    -1,
+    0,
+    -1,
+    3,
+    -1,
+    -3,
+    2,
+    1,
+    -2,
+    3,
+    -2,
+    1,
+    0,
+    0,
+    1,
+    -3,
+    2,
+    -1,
+    2,
+    -1,
+    -1,
+    0,
+    1,
+    0,
+    0,
+    1,
+    -2,
+    0,
+    -1,
+    3,
+    -2,
+    -1,
+    2,
+    0,
+    0,
+    -1,
+    1,
+    -2,
+    3,
+    -1,
+    -1,
+    -1,
+    1,
+    2,
+    -2,
+    1,
+    -2,
+    0,
+    0,
+    0,
+    1,
+    0,
+    0,
+    -1,
+    0,
+    1,
+    -1,
+    1,
+    0,
+    -1,
+    -1,
+    1,
+    1,
+    0,
+    0,
+    -2,
+    5,
+    -1,
+    -2,
+    2,
+    -2,
+    0,
+    1,
+    0,
+    -1,
+    0,
+    1,
+    0,
+    -1,
+    2,
+    -3,
+    1,
+    -1,
+    1,
+    1,
+    1,
+    -2,
+    0,
+    2,
+    -3,
+    3,
+    -3,
+    1,
+    1,
+    -1,
+    1,
+    -1,
+    -1,
+    2,
+    0,
+    -2,
+    1,
+    2,
+    -1,
+    0,
+    0,
+    0,
+    1,
+    -2,
+    -1,
+    0,
+    1,
+    1,
+    0,
+    -1,
+    -1,
+    3,
+    -2,
+    -2,
+    3,
+    0,
+    0,
+    0,
+    -2,
+    3,
+    -1,
+    0,
+    -1,
+    1,
+    1,
+    0,
+    -3,
+    2,
+    -1,
+    2,
+    -2,
+    -2,
+    2,
+    0,
+    1,
+    1,
+    0,
+    -4,
+    2,
+    1,
+    0,
+    1,
+    -3,
+    -1,
+    3,
+    0,
+    0,
+    -1,
+    0,
+    2,
+    -2,
+    2,
+    -2,
+    1,
+    -1,
+    1,
+    0,
+    0,
+    -2,
+    3,
+    -4,
+    1,
+    3,
+    -1,
+    -1,
+    1,
+    0,
+    0,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    0,
+    0,
+    0,
+    0,
+    2,
+    -2,
+    2,
+    -2,
+    0,
+    1,
+    -2,
+    3,
+    -1,
+    1,
+    -1,
+    0,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    -2,
+    2,
+    1,
+    -3,
+    1,
+    -1,
+    1,
+    -1,
+    -1,
+    2,
+    0,
+    1,
+    0,
+    -2,
+    0,
+    1,
+    1,
+    0,
+    -1,
+    0,
+    0,
+    0,
+    -2,
+    2,
+    -2,
+    1,
+    2,
+    -1,
+    0,
+    -1,
+    2,
+    -3,
+    4,
+    -2,
+    0,
+    -1,
+    0,
+    0,
+    0,
+    0,
+    0,
+    1,
+    1,
+    -3,
+    1,
+    1,
+    -2,
+    2,
+    -1,
+    0,
+    1,
+    -2,
+    2,
+    -2,
+    2,
+    -1,
+    -1,
+    2,
+    -1,
+    1,
+    -2,
+    2,
+    -2,
+    2,
+    1,
+    -3,
+    -1,
+    2,
+    3,
+    -2,
+    1,
+    -1,
+    -1,
+    1,
+    -1,
+    1,
+    1,
+    -3,
+    3,
+    1,
+    -4,
+    0,
+    1,
+    2,
+    0,
+    -2,
+    0,
+    0,
+    1,
+    0,
+    1,
+    -3,
+    2,
+    0,
+    0,
+    1,
+    -3,
+    1,
+    2,
+    -2,
+    2,
+    -1,
+    0,
+    0,
+    -1,
+    -1,
+    2,
+    -2,
+    1,
+    1,
+    0,
+    1,
+    -2,
+    0,
+    0,
+    1,
+    0,
+    0,
+    -1,
+    1,
+    -2,
+    1,
+    1,
+    0,
+    1,
+    -1,
+    0,
+    -1,
+    1,
+    -1,
+    -1,
+    2,
+    0,
+    1,
+    -2,
+    1,
+    0,
+    -1,
+    1,
+    1,
+    -3,
+    2,
+    0,
+    -1,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    -2,
+    2,
+    -1,
+    1,
+    0,
+    -1,
+    -1,
+    1,
+    1,
+    1,
+    -3,
+    2,
+    -1,
+    1,
+    -2,
+    2,
+    1,
+    -2,
+    2,
+    -2,
+    1,
+    -1,
+    1,
+    1,
+    -2,
+    1,
+    0,
+    1,
+    -2,
+    1,
+    -2,
+    2,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    -3,
+    1,
+    0,
+    1,
+    1,
+    0,
+    -2,
+    -1,
+    3,
+    -1,
+    0,
+    1,
+    0,
+    -2,
+    0,
+    1,
+    0,
+    -1,
+    0,
+    2,
+    -2,
+    -2,
+    1,
+    2,
+    2,
+    -2,
+    -1,
+    -1,
+    3,
+    0,
+    0,
+    -2,
+    -1,
+    1,
+    0,
+    -1,
+    2,
+    1,
+    -2,
+    1,
+    -1,
+    0,
+    0,
+    -1,
+    3,
+    0,
+    1,
+    0,
+    0,
+    -3,
+    0,
+    1,
+    1,
+    0,
+    0,
+    -2,
+    -1,
+    3,
+    0,
+    -1,
+    -1,
+    1,
+    -1,
+    0,
+    1,
+    -1,
+    -1,
+    0,
+    1,
+    2,
+    -2,
+    2,
+    -1,
+    0,
+    -1,
+    1,
+    -2,
+    0,
+    1,
+    0,
+    3,
+    -5,
+    3,
+    -1,
+    0,
+    0,
+    3,
+    -2,
+    0,
+    1,
+    -1,
+    0,
+    0,
+    -2,
+    1,
+    1,
+    0,
+    0,
+    -1,
+    0,
+    2,
+    0,
+    -2,
+    -1,
+    2,
+    -1,
+    0,
+    1,
+    -1,
+    0,
+    3,
+    -2,
+    -1,
+    -1,
+    0,
+    3,
+    -1,
+    -2,
+    0,
+    2,
+    -2,
+    2,
+    0,
+    1,
+    -3,
+    1,
+    -1,
+    2,
+    0,
+    0,
+    -1,
+    2,
+    -1,
+    1,
+    -1,
+    -1,
+    1,
+    0,
+    1,
+    -2,
+    0,
+    0,
+    0,
+    1,
+    0,
+    0,
+    0,
+    0,
+    0,
+    -2,
+    2,
+    -2,
+    0,
+    3,
+    -2,
+    2,
+    -2,
+    1,
+    1,
+    0,
+    -2,
+    -1,
+    3,
+    -2,
+    0,
+    0,
+    0,
+    0,
+    1,
+    -2,
+    2,
+    -2,
+    3,
+    -1,
+    -1,
+    -1,
+    1,
+    -2,
+    5,
+    -3,
+    1,
+    -1,
+    3,
+    -3,
+    0,
+    1,
+    -1,
+    1,
+    -1,
+    -2,
+    4,
+    -2,
+    1,
+    1,
+    -1,
+    1,
+    -2,
+    1,
+    -2,
+    3,
+    -1,
+    0,
+    0,
+    0,
+    1,
+    -1,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    0,
+    -1,
+    1,
+    1,
+    -1,
+    0,
+    0,
+    -1,
+    1,
+    0,
+    -2,
+    2,
+    0,
+    -1,
+    0,
+    2,
+    -1,
+    0,
+    -2,
+    1,
+    0,
+    0,
+    -1,
+    2,
+    -3,
+    3,
+    1,
+    -1,
+    -1,
+    -2,
+    2,
+    0,
+    1,
+    -1,
+    1,
+    -3,
+    1,
+    1,
+    2,
+    -2,
+    -1,
+    -1,
+    2,
+    -2,
+    0,
+    3,
+    -2,
+    1,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    2,
+    -1,
+    0,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    1,
+    -1,
+    0,
+    1,
+    0,
+    -2,
+    2,
+    -2,
+    1,
+    -1,
+    0,
+    1,
+    -1,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    1,
+    -2,
+    -1,
+    3,
+    0,
+    -1,
+    -1,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    0,
+    0,
+    -1,
+    -3,
+    0,
+    2,
+    0,
+    0,
+    0,
+    -1,
+    0,
+    1,
+    1,
+    -2,
+    4,
+    -4,
+    0,
+    0,
+    0,
+    1,
+    -1,
+    1,
+    1,
+    0,
+    -2,
+    2,
+    0,
+    1,
+    0,
+    -2,
+    0,
+    -1,
+    0,
+    2,
+    0,
+    -2,
+    3,
+    -2,
+    0,
+    1,
+    -2,
+    0,
+    1,
+    0,
+    3,
+    -4,
+    1,
+    -1,
+    2,
+    -2,
+    1,
+    -1,
+    0,
+    1,
+    1,
+    -2,
+    4,
+    -2,
+    1,
+    -1,
+    -1,
+    0,
+    2,
+    -2,
+    1,
+    -1,
+    1,
+    2,
+    -4,
+    3,
+    -1,
+    1,
+    0,
+    -1,
+    1,
+    -2,
+    1,
+    0,
+    -2,
+    3,
+    -2,
+    2,
+    -2,
+    0,
+    3,
+    -2,
+    0,
+    1,
+    -2,
+    2,
+    -1,
+    -1,
+    1,
+    1,
+    -1,
+    1,
+    -1,
+    0,
+    0,
+    0,
+    -1,
+    2,
+    -1,
+    0,
+    -2,
+    3,
+    -2,
+    0,
+    1,
+    0,
+    0,
+    0,
+    0,
+    -2,
+    3,
+    -1,
+    -1,
+    1,
+    -1,
+    1,
+    0,
+    0,
+    0,
+    -1,
+    1,
+    -2,
+    1,
+    0,
+    0,
+    -1,
+    3,
+    -1,
+    1,
+    -1,
+    -2,
+    1,
+    0,
+    1,
+    0,
+    -2,
+    0,
+    3,
+    -3,
+    2,
+    2,
+    -3,
+    2,
+    -1,
+    0,
+    -1,
+    -1,
+    2,
+    -1,
+    1,
+    0,
+    -2,
+    2,
+    -1,
+    3,
+    -2,
+    -1,
+    0,
+    -1,
+    1,
+    2,
+    -3,
+    1,
+    2,
+    -1,
+    0,
+    -1,
+    1,
+    0,
+    0,
+    -1,
+    0,
+    -2,
+    3,
+    -2,
+    2,
+    -1,
+    -1,
+    2,
+    -1,
+    1,
+    0,
+    0,
+    2,
+    -3,
+    -1,
+    0,
+    1,
+    0,
+    -2,
+    2,
+    0,
+    0,
+    0,
+    -1,
+    3,
+    -2,
+    -1,
+    2,
+    -3,
+    2,
+    0,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    -1,
+    1,
+    0,
+    -1,
+    0,
+    0,
+    0,
+    1,
+    -2,
+    1,
+    -1,
+    0,
+    4,
+    -2,
+    -1,
+    0,
+    1,
+    0,
+    -1,
+    -1,
+    2,
+    1,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    -2,
+    2,
+    0,
+    -1,
+    0,
+    1,
+    -1,
+    2,
+    -2,
+    1,
+    0,
+    -3,
+    3,
+    0,
+    -1,
+    -1,
+    2,
+    0,
+    0,
+    -2,
+    2,
+    1,
+    -1,
+    -1,
+    1,
+    0,
+    -2,
+    1,
+    1,
+    -1,
+    2,
+    -4,
+    4,
+    -1,
+    -1,
+    1,
+    -1,
+    1,
+    0,
+    0,
+    -1,
+    -1,
+    2,
+    -2,
+    0,
+    1,
+    1,
+    -1,
+    -1,
+    1,
+    1,
+    2,
+    -2,
+    -2,
+    1,
+    0,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    2,
+    -1,
+    2,
+    -3,
+    2,
+    0,
+    -1,
+    0,
+    -1,
+    0,
+    2,
+    -2,
+    -1,
+    2,
+    -1,
+    2,
+    0,
+    -1,
+    0,
+    1,
+    -1,
+    0,
+    1,
+    -1,
+    0,
+    -1,
+    -2,
+    2,
+    2,
+    -3,
+    2,
+    0,
+    -2,
+    4,
+    -2,
+    0,
+    -1,
+    -1,
+    0,
+    2,
+    -2,
+    2,
+    0,
+    0,
+    -1,
+    1,
+    -1,
+    0,
+    0,
+    0,
+    -1,
+    3,
+    -2,
+    2,
+    -1,
+    0,
+    0,
+    -1,
+    -1,
+    1,
+    0,
+    0,
+    1,
+    0,
+    0,
+    0,
+    -1,
+    1,
+    -2,
+    3,
+    -2,
+    1,
+    0,
+    0,
+    -1,
+    -1,
+    0,
+    2,
+    0,
+    -1,
+    0,
+    0,
+    0,
+    2,
+    -4,
+    5,
+    -2,
+    2,
+    -3,
+    2,
+    -3,
+    3,
+    0,
+    -3,
+    3,
+    -2,
+    3,
+    -1,
+    -1,
+    -1,
+    1,
+    1,
+    -1,
+    -2,
+    3,
+    -2,
+    1,
+    0,
+    -2,
+    0,
+    3,
+    -2,
+    1,
+    1,
+    -2,
+    0,
+    -2,
+    2,
+    1,
+    1,
+    -2,
+    1,
+    1,
+    -2,
+    1,
+    -1,
+    0,
+    1,
+    0,
+    0,
+    2,
+    -3,
+    1,
+    0,
+    -1,
+    1,
+    0,
+    -1,
+    3,
+    -2,
+    -1,
+    1,
+    -2,
+    3,
+    -1,
+    -1,
+    1,
+    -2,
+    2,
+    2,
+    -2,
+    -2,
+    3,
+    -2,
+    0,
+    1,
+    0,
+    -1,
+    1,
+    -2,
+    2,
+    -3,
+    3,
+    0,
+    -1,
+    1,
+    1,
+    -2,
+    -1,
+    1,
+    1,
+    0,
+    1,
+    1,
+    -2,
+    -1,
+    1,
+    1,
+    -1,
+    1,
+    -1,
+    1,
+    0,
+    0,
+    -1,
+    -1,
+    1,
+    -1,
+    1,
+    1,
+    -1,
+    1,
+    -1
+};
