+++ passphraseToId2.cl
@@ -2348,7 +2348,7 @@ void idToByteAccount(ulong accountId, uchar * out) {
     }
 }
 
-#define PASSPHRASE_MAX_LENGTH 135
+#define PASSPHRASE_MAX_LENGTH 143
 
 struct PASSPHRASE {
     uchar offset;
